requirements.txt
flask
google-cloud-texttospeech
pydub
openai
python-dotenv